my-ai-assistant/backend/requirements.txt
zhipuai>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
chromadb>=0.4.0
pydantic>=2.5.0
sqlalchemy>=2.0.0
python-multipart>=0.0.6
